requirements.txt
Django==2.2
coreapi==2.3.3
djangorestframework==3.11.0
